blueprints/datalake-warehouse-c360/project/requirements.txt
aws-cdk-lib==2.130.0
aws-cdk.aws-glue-alpha==2.130.0a0
aws-cdk-aws-lambda-python-alpha==2.130.0a0
aws-cdk-aws-redshift-alpha==2.130.0a0
aws-cdk-aws-msk-alpha==2.130.0a0
cdk-nag==2.28.18
constructs==10.1.244
boto3==1.26.64
botocore==1.29.64
Faker==16.6.1
aws_solutions_constructs.aws_lambda_opensearch
numpy
pandas
nltk
pyyaml
ray
jsonpickle 
PyMySQL
aws_lambda_powertools
aws-cdk.aws-cloud9-alpha
opensearch-py
dbt-redshift 
apache-airflow==2.9.0
apache-airflow-providers-salesforce==5.1.0
apache-airflow-providers-amazon==6.0.0
apache-airflow-providers-ssh==3.2.0
apache-airflow-providers-common-sql==1.2.0
wheel 